Extract P2P DAG tests from production modules
